# Sped up covariance estimate via numba