--- a/arch/covariance/kernel.py
+++ b/arch/covariance/kernel.py
@@ -140,7 +140,6 @@ class CovarianceEstimate:
 @jit(nopython=True)
 def _cov_jit(df, k, num_weights, w, x):
     oss = np.zeros((k, k))
-    x = np.ascontiguousarray(x)
     for i in range(1, num_weights):
         oss += w[i] * (x[i:].T @ x[:-i]) / df
     return oss
@@ -402,7 +401,7 @@ class CovarianceEstimator(ABC):
         --------
         CovarianceEstimate
         """
-        x = np.asarray(self._x)
+        x = np.ascontiguousarray(self._x)
         k = x.shape[1]
         df = self._df
         sr = x.T @ x / df