Benchmarked using: `py.test arch/tests/unitroot/test_fmols_ccr.py`

Before:
```bash
============================= fixture duration top ============================== 
total          name              num   avg            min
0:00:00.080000             trend  1728        0:00:00 0:00:00
0:00:00.062000              diff  1728        0:00:00 0:00:00
0:00:00.032000            kernel  1728        0:00:00 0:00:00
0:00:00.031000         force_int  1728        0:00:00 0:00:00
0:00:00.016000   trivariate_data     2 0:00:00.008000 0:00:00
0:00:00.016000           x_trend  1728        0:00:00 0:00:00
0:00:00.016000         bandwidth  1728        0:00:00 0:00:00
0:00:00.253000       grand total 10426        0:00:00 0:00:00
============================ test call duration top ============================= 
total          name              num   avg            min
0:00:06.062000  test_fmols_smoke   864        0:00:00 0:00:00
0:00:06.001000    test_ccr_smoke   864        0:00:00 0:00:00
0:00:00.093000   test_ccr_eviews    26        0:00:00 0:00:00
0:00:00.078000 test_fmols_eviews    26        0:00:00 0:00:00
0:00:12.234000       grand total  1784        0:00:00 0:00:00
============================ test setup duration top ============================ 
total          name              num   avg            min
0:00:00.377000  test_fmols_smoke   864        0:00:00 0:00:00
0:00:00.295000    test_ccr_smoke   864        0:00:00 0:00:00
0:00:00.672000       grand total  1784        0:00:00 0:00:00
========================== test teardown duration top =========================== 
total          name              num   avg            min
0:00:00.142000  test_fmols_smoke   864        0:00:00 0:00:00
0:00:00.124000    test_ccr_smoke   864        0:00:00 0:00:00
0:00:00.266000       grand total  1784        0:00:00 0:00:00
============================= 1784 passed in 14.37s ============================= 
```

After:
```
============================= fixture duration top ============================== 
total          name              num   avg            min
0:00:00.063000             trend  1728        0:00:00 0:00:00
0:00:00.048000           x_trend  1728        0:00:00 0:00:00
0:00:00.032000         force_int  1728        0:00:00 0:00:00
0:00:00.016000   trivariate_data     2 0:00:00.008000 0:00:00
0:00:00.016000            kernel  1728        0:00:00 0:00:00
0:00:00.015000         bandwidth  1728        0:00:00 0:00:00
0:00:00.190000       grand total 10426        0:00:00 0:00:00
============================ test call duration top ============================= 
total          name              num   avg            min
0:00:04.528000  test_fmols_smoke   864        0:00:00 0:00:00
0:00:03.748000    test_ccr_smoke   864        0:00:00 0:00:00
0:00:00.094000 test_fmols_eviews    26        0:00:00 0:00:00
0:00:00.094000   test_ccr_eviews    26        0:00:00 0:00:00
0:00:08.464000       grand total  1784        0:00:00 0:00:00
============================ test setup duration top ============================ 
total          name              num   avg            min
0:00:00.392000    test_ccr_smoke   864        0:00:00 0:00:00
0:00:00.329000  test_fmols_smoke   864        0:00:00 0:00:00
0:00:00.721000       grand total  1784        0:00:00 0:00:00
========================== test teardown duration top =========================== 
total          name              num   avg            min
0:00:00.109000  test_fmols_smoke   864        0:00:00 0:00:00
0:00:00.064000    test_ccr_smoke   864        0:00:00 0:00:00
0:00:00.016000   test_ccr_eviews    26        0:00:00 0:00:00
0:00:00.189000       grand total  1784        0:00:00 0:00:00
============================= 1784 passed in 10.34s ============================= 
```

Nothing too dramatic but a useful performance improvement :)

I'm still looking at other bottlenecks right now so depending on your preference I can create separate pull requests or a single large one :)