@wolph 2023-10-11
I actually ran `isort`, `black` and `flake8` and none of those show issues with my code.

I'm guessing there's been a `black` update this is the diff that `black` is producing now:
```
$ git diff
diff --git a/arch/univariate/base.py b/arch/univariate/base.py
index 4847000f8..f59fd7c1c 100644
--- a/arch/univariate/base.py
+++ b/arch/univariate/base.py
@@ -1189,10 +1189,7 @@ class ARCHModelFixedResult(_SummaryRepr):

         stubs = list(self._names)
         header = ["coef"]
-        param_table_data = [
-            [format_float_fixed(param, 10, 4)]
-            for param in self.params
-        ]
+        param_table_data = [[format_float_fixed(param, 10, 4)] for param in self.params]

         mc = self.model.num_params
         vc = self.model.volatility.num_params
```

I can apply it if you want to, but it's not really part of this PR :)

The other issue is (was, I just pushed an update) that the signature for the `jit` function is incorrect. It doesn't allow usage as a decorator with arguments