@bashtage 2023-10-12
To get the decorator to work, I thnk you can use 


```python
try:
    if DISABLE_NUMBA:
        raise ImportError

    from numba import jit

    jit = functools.partial(jit, nopython=True)

except ImportError:

    def jit(
        *args: Any,
        **kwargs: Any,
    ) -> Any:
        if args and callable(args[0]):
            func = args[0]

        def wrap(func):
            def wrapper(*args: Any, **kwargs: Any) -> Callable[..., Any]:
                import warnings

                warnings.warn(performance_warning, PerformanceWarning)
                return func(*args, **kwargs)

            return wrapper

        return wrap
```