@bashtage 2023-10-12
I pushed an improved version of the fix.